+++ App-Project/src/main/java/com/springbootcamp/springsecurity/domain/Role.java
@@ -11,7 +11,7 @@ import java.util.List;
 public class Role {
     @Id
     //@GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     private String authority;
 
 
@@ -32,16 +32,16 @@ public class Role {
 
     }
 
-    public Role(Integer id, String authority) {
+    public Role(Long id, String authority) {
         this.id = id;
         this.authority = authority;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
