+++ App-Project/src/main/java/com/springbootcamp/springsecurity/repository/OrderRepository.java
@@ -3,5 +3,5 @@ package com.springbootcamp.springsecurity.repository;
 import com.springbootcamp.springsecurity.domain.order.Orders;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderRepository extends CrudRepository<Orders,Integer> {
+public interface OrderRepository extends CrudRepository<Orders,Long> {
 }
