+++ App-Project/src/main/java/com/springbootcamp/springsecurity/repository/ProductRepository.java
@@ -3,6 +3,6 @@ package com.springbootcamp.springsecurity.repository;
 import com.springbootcamp.springsecurity.domain.product.Products;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Products,Integer> {
+public interface ProductRepository extends CrudRepository<Products,Long> {
     Products findByName(String name);
 }
