+++ App-Project/src/main/java/com/springbootcamp/springsecurity/repository/SellerRepository.java
@@ -3,6 +3,6 @@ package com.springbootcamp.springsecurity.repository;
 import com.springbootcamp.springsecurity.domain.user.Sellers;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SellerRepository extends CrudRepository<Sellers,Integer> {
+public interface SellerRepository extends CrudRepository<Sellers,Long> {
     Sellers findByEmail(String string);
 }
